ParameterContentNegotiationStrategy uses MediaTypeFactory
Issue: SPR-15649
