+++ spring-web/src/main/java/org/springframework/web/accept/ContentNegotiationManager.java
@@ -44,7 +44,7 @@ import org.springframework.web.context.request.NativeWebRequest;
  */
 public class ContentNegotiationManager implements ContentNegotiationStrategy, MediaTypeFileExtensionResolver {
 
-	private static final List<MediaType> MEDIA_TYPE_ALL = Collections.<MediaType>singletonList(MediaType.ALL);
+	private static final List<MediaType> MEDIA_TYPE_ALL = Collections.singletonList(MediaType.ALL);
 
 
 	private final List<ContentNegotiationStrategy> strategies = new ArrayList<>();
