+++ spring-web/src/main/java/org/springframework/web/accept/ParameterContentNegotiationStrategy.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2016 the original author or authors.
+ * Copyright 2002-201 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,24 +18,23 @@ package org.springframework.web.accept;
 
 import java.util.Map;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-
 import org.springframework.http.MediaType;
 import org.springframework.util.Assert;
-import org.springframework.web.HttpMediaTypeNotAcceptableException;
 import org.springframework.web.context.request.NativeWebRequest;
 
 /**
- * A {@code ContentNegotiationStrategy} that resolves a query parameter to a key
- * to be used to look up a media type. The default parameter name is {@code format}.
+ * Strategy that resolves the requested content type from a query parameter.
+ * The default query parameter name is {@literal "format"}.
+ *
+ * <p>You can register static mappings between keys (i.e. the expected value of
+ * the query parameter) and MediaType's via {@link #addMapping(String, MediaType)}.
+ * As of 5.0 this strategy also supports dynamic lookups of keys via
+ * {@link org.springframework.http.MediaTypeFactory#getMediaType}.
  *
  * @author Rossen Stoyanchev
  * @since 3.2
  */
 public class ParameterContentNegotiationStrategy extends AbstractMappingContentNegotiationStrategy {
-
-	private static final Log logger = LogFactory.getLog(ParameterContentNegotiationStrategy.class);
 
 	private String parameterName = "format";
 
@@ -65,21 +64,6 @@ public class ParameterContentNegotiationStrategy extends AbstractMappingContentN
 	@Override
 	protected String getMediaTypeKey(NativeWebRequest request) {
 		return request.getParameter(getParameterName());
-	}
-
-	@Override
-	protected void handleMatch(String mediaTypeKey, MediaType mediaType) {
-		if (logger.isDebugEnabled()) {
-			logger.debug("Requested media type: '" + mediaType + "' based on '" +
-					getParameterName() + "'='" + mediaTypeKey + "'");
-		}
-	}
-
-	@Override
-	protected MediaType handleNoMatch(NativeWebRequest request, String key)
-			throws HttpMediaTypeNotAcceptableException {
-
-		throw new HttpMediaTypeNotAcceptableException(getAllMediaTypes());
 	}
 
 }
