+++ spring-web/src/main/java/org/springframework/web/accept/PathExtensionContentNegotiationStrategy.java
@@ -18,12 +18,7 @@ package org.springframework.web.accept;
 
 import java.util.Locale;
 import java.util.Map;
-import java.util.Optional;
-
 import javax.servlet.http.HttpServletRequest;
-
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 
 import org.springframework.core.io.Resource;
 import org.springframework.http.MediaType;
@@ -31,7 +26,6 @@ import org.springframework.http.MediaTypeFactory;
 import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
 import org.springframework.util.StringUtils;
-import org.springframework.web.HttpMediaTypeNotAcceptableException;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.util.UriUtils;
 import org.springframework.web.util.UrlPathHelper;
@@ -49,13 +43,7 @@ import org.springframework.web.util.UrlPathHelper;
  */
 public class PathExtensionContentNegotiationStrategy extends AbstractMappingContentNegotiationStrategy {
 
-	private static final Log logger = LogFactory.getLog(PathExtensionContentNegotiationStrategy.class);
-
 	private UrlPathHelper urlPathHelper = new UrlPathHelper();
-
-	private boolean useRegisteredExtensionsOnly = false;
-
-	private boolean ignoreUnknownExtensions = true;
 
 
 	/**
@@ -71,6 +59,8 @@ public class PathExtensionContentNegotiationStrategy extends AbstractMappingCont
 	 */
 	public PathExtensionContentNegotiationStrategy(@Nullable Map<String, MediaType> mediaTypes) {
 		super(mediaTypes);
+		setUseRegisteredExtensionsOnly(false);
+		setIgnoreUnknownExtensions(true);
 		this.urlPathHelper.setUrlDecode(false);
 	}
 
@@ -92,25 +82,6 @@ public class PathExtensionContentNegotiationStrategy extends AbstractMappingCont
 		setUseRegisteredExtensionsOnly(!useJaf);
 	}
 
-	/**
-	 * Whether to only use the registered mappings to look up file extensions, or also refer to
-	 * defaults.
-	 * <p>By default this is set to {@code false}, meaning that defaults are used.
-	 */
-	public void setUseRegisteredExtensionsOnly(boolean useRegisteredExtensionsOnly) {
-		this.useRegisteredExtensionsOnly = useRegisteredExtensionsOnly;
-	}
-
-	/**
-	 * Whether to ignore requests with unknown file extension. Setting this to
-	 * {@code false} results in {@code HttpMediaTypeNotAcceptableException}.
-	 * <p>By default this is set to {@code true}.
-	 */
-	public void setIgnoreUnknownExtensions(boolean ignoreUnknownExtensions) {
-		this.ignoreUnknownExtensions = ignoreUnknownExtensions;
-	}
-
-
 	@Override
 	protected String getMediaTypeKey(NativeWebRequest webRequest) {
 		HttpServletRequest request = webRequest.getNativeRequest(HttpServletRequest.class);
@@ -121,22 +92,6 @@ public class PathExtensionContentNegotiationStrategy extends AbstractMappingCont
 		String path = this.urlPathHelper.getLookupPathForRequest(request);
 		String extension = UriUtils.extractFileExtension(path);
 		return (StringUtils.hasText(extension) ? extension.toLowerCase(Locale.ENGLISH) : null);
-	}
-
-	@Override
-	protected MediaType handleNoMatch(NativeWebRequest webRequest, String extension)
-			throws HttpMediaTypeNotAcceptableException {
-
-		if (!this.useRegisteredExtensionsOnly) {
-			Optional<MediaType> mediaType = MediaTypeFactory.getMediaType("file." + extension);
-			if (mediaType.isPresent()) {
-				return mediaType.get();
-			}
-		}
-		if (this.ignoreUnknownExtensions) {
-			return null;
-		}
-		throw new HttpMediaTypeNotAcceptableException(getAllMediaTypes());
 	}
 
 	/**
